% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/network_simulation.R
\name{add.edgecov}
\alias{add.edgecov}
\title{add edge-level covariates to a network data frame}
\usage{
add.edgecov(net.df)
}
\arguments{
\item{net.df}{a network data frame}
}
\value{
a data frame
}
\description{
A one-sentence description of what the function does.
}
