% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/network_simulation.R
\name{count.infected.neighbors}
\alias{count.infected.neighbors}
\title{count the number of infected neighbors for each individual at time k}
\usage{
\method{count}{infected.neighbors}(net.df, k)
}
\arguments{
\item{net.df}{a network data frame}

\item{k}{a positive integer, the time point at which to count infected neighbors}
}
\value{
a numeric vector, the counts of infected neighbors
}
\description{
A one-sentence description of what the function does.
}
