% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/di_stergm.R
\name{di.stergm.fit}
\alias{di.stergm.fit}
\title{fit dyad independence STERGM}
\usage{
di.stergm.fit(net.df, theta.p.start, theta.m.start)
}
\arguments{
\item{net.df}{a network data frame}

\item{theta.p.start}{a numeric vector, starting value for formation model parameters}

\item{theta.m.start}{a numeric vector, starting value for the persistence model parameters}
}
\value{
a numeric matrix, the Hessian value
}
\description{
A one-sentence description of what the function does.
}
