% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/di_stergm.R
\name{di.stergm.loglik}
\alias{di.stergm.loglik}
\title{log-likelihood for a dyad independence STERGM}
\usage{
di.stergm.loglik(net.df, theta.p, theta.m)
}
\arguments{
\item{net.df}{a network data frame}

\item{theta.p}{a numeric vector, the formation model parameters}

\item{theta.m}{a numeric vector, the persistence model parameters}
}
\value{
a number, the log-likelihood value
}
\description{
A one-sentence description of what the function does.
}
