% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/network_simulation.R
\name{g}
\alias{g}
\title{get infection probability given number of infected neighbors}
\usage{
g(s, H0Y, EY = NULL, g.shift = 8.5, g.scale = 0.8)
}
\arguments{
\item{s}{a numeric vector, counts of infected neighbors}

\item{H0Y}{a logical indicator for whether H_0^Y is true}

\item{EY}{a number in \link{0,1}, the (constant) probability of infection if \code{H0Y = TRUE}}

\item{g.shift}{a number, the shift parameter in \code{g} if \code{H0Y = FALSE}}

\item{g.scale}{a number, the scale parameter in \code{g} if \code{H0Y = FALSE}}
}
\value{
a numeric vector, the infection probabilities
}
\description{
A one-sentence description of what the function does.
}
