% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/randomization_tests.R
\name{get.rho.adj}
\alias{get.rho.adj}
\title{get bootstrap adjusted p-value testing the nonsharp null}
\usage{
get.rho.adj(
  net.df,
  Z,
  randomize,
  arguments,
  B,
  R,
  theta.p,
  theta.m,
  rho.plug,
  pb = F
)
}
\arguments{
\item{net.df}{a network data frame}

\item{randomize}{a function, the randomization mechanism}

\item{arguments}{a list, arguments for the randomization mechanism}

\item{B}{a positive integer, the number of MC replicates}

\item{R}{a positive integer, the number of bootstrap replicates}

\item{theta.p}{a numeric vector, the formation model parameters}

\item{theta.m}{a numeric vector, the persistence model parameters}

\item{rho.plug}{a numeric vector, the plug-in p-values testing the nonsharp null}

\item{pb}{a logical indicator for whether to print a progress bar, default is \code{TRUE}}
}
\value{
a numeric vector, p-values
}
\description{
A one-sentence description of what the function does.
}
