% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/randomization_tests.R
\name{get.rho}
\alias{get.rho}
\title{compute a p-value given observed test statistic and null distribution}
\usage{
get.rho(T.obs, T.distr)
}
\arguments{
\item{T.obs}{a numeric vector, observed values of test statistics}

\item{T.distr}{a numeric matrix, the null distributions of test statistics}
}
\value{
a numeric vector, p-values
}
\description{
A one-sentence description of what the function does.
}
