% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_stats.R
\name{get.T.all}
\alias{get.T.all}
\title{get test statistics using 12 different definitions of transmission event}
\usage{
get.T.all(net.df, Z)
}
\arguments{
\item{net.df}{a network data frame}

\item{theta.p}{a binary numeric vector, randomization assignments}
}
\value{
a numeric vector, the values of the 12 test statistics
}
\description{
A one-sentence description of what the function does.
}
