% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/randomization_tests.R
\name{get.T.sharp}
\alias{get.T.sharp}
\title{simulate the null distribution of test statistics under the sharp null}
\usage{
get.T.sharp(net.df, randomize, arguments, B, pb = T)
}
\arguments{
\item{net.df}{a network data frame}

\item{randomize}{a function, the randomization mechanism}

\item{arguments}{a list, arguments for the randomization mechanism}

\item{B}{a positive integer, the number of MC replicates}

\item{pb}{a logical indicator for whether to print a progress bar, default is \code{TRUE}}
}
\value{
a numeric vector, p-values
}
\description{
A one-sentence description of what the function does.
}
