% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/network_simulation.R
\name{initialize.net.df}
\alias{initialize.net.df}
\title{initialize a network data frame}
\usage{
initialize.net.df(n, tau)
}
\arguments{
\item{n}{a positive integer, the sample size}

\item{tau}{a positive integer, the number of time points}
}
\value{
a data frame
}
\description{
A one-sentence description of what the function does.
}
