% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/randomization_mechanisms.R
\name{randomize.bernoulli}
\alias{randomize.bernoulli}
\title{Bernoulli randomization}
\usage{
randomize.bernoulli(arguments)
}
\arguments{
\item{arguments}{a list with \code{n} (sample size) and \code{p} (probability of treatment)}
}
\value{
a numeric vector, the randomization assignments
}
\description{
A one-sentence description of what the function does.
}
