% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/network_simulation.R
\name{simulate.fixedcov}
\alias{simulate.fixedcov}
\title{simulate data according to a dyad independence STERGM with prespecified covariates}
\usage{
\method{simulate}{fixedcov}(net.df, theta.p, theta.m, print.progress = TRUE)
}
\arguments{
\item{net.df}{an initialized network data frame}

\item{theta.p}{a numeric vector, the formation model parameters}

\item{theta.m}{a numeric vector, the persistence model parameters}

\item{print.progress}{a logical indicator for whether to print progress, default is \code{TRUE}}
}
\value{
a network data frame
}
\description{
A one-sentence description of what the function does.
}
