% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/network_simulation.R
\name{simulate.newcov}
\alias{simulate.newcov}
\title{simulate data according to a dyad independence STERGM with new randomizations and infection statuses}
\usage{
\method{simulate}{newcov}(
  n,
  tau,
  theta.p,
  theta.m,
  randomize,
  arguments,
  H0Y,
  EY = NULL,
  g.shift = NULL,
  g.scale = NULL,
  print.progress = TRUE
)
}
\arguments{
\item{theta.p}{a numeric vector, the formation model parameters}

\item{theta.m}{a numeric vector, the persistence model parameters}

\item{randomize}{a function, the randomization mechanism}

\item{arguments}{a list, arguments for the randomization mechanism}

\item{H0Y}{a logical indicator for whether H_0^Y is true}

\item{EY}{a number in \link{0,1}, the (constant) probability of infection if \code{H0Y = TRUE}}

\item{g.shift}{a number, the shift parameter in \code{g} if \code{H0Y = FALSE}}

\item{g.scale}{a number, the scale parameter in \code{g} if \code{H0Y = FALSE}}

\item{print.progress}{a logical indicator for whether to print progress, default is \code{TRUE}}

\item{net.df}{an initialized network data frame}
}
\value{
a list with randomization assignments \code{Z} and a network data frame \code{net.df}
}
\description{
A one-sentence description of what the function does.
}
